Big Refactor - move all videostreaming-related code into its own subdirectory, and generify by used librarie(s)
