+++ app/videostreaming/android/qandroidmediaplayer.cpp
