+++ app/videostreaming/android/qandroidmediaplayer.h
