+++ app/videostreaming/android/videostreamingandroid.pri
